Rewrite list_open_prs.py and limit slack mentions
Rewrite list_open_prs.py to use a more object oriented approach to make
code easier to read.

Check who should be pinged on slack based on who can act on the PR:

Author is pinged if the PR title starts with "WIP" or if there are no
pending reviews.

Reviewers are pinged if the PR has been updated since they last
commented or requested changes or if they haven't reviewed yet.